+++ sample/src/main/java/com/synerise/sdk/sample/ui/dev/apiAdapter/fragments/promotion/ClientGetPromotionsFragment.java
@@ -32,7 +32,7 @@ public class ClientGetPromotionsFragment extends BaseDevFragment {
     private IDataApiCall<PromotionResponse> apiCall;
     private RadioButton includeButton;
     private CheckBox activeBox, assignedBox, redeemedBox;
-    private CheckBox generalBox, customBox, membersOnlyBox, includeMetaBox, checkGlobalActivationLimitBox;
+    private CheckBox generalBox, customBox, membersOnlyBox, includeMetaBox, checkGlobalActivationLimitBox, checkBoxincludeVouchers;
     private TextInputLayout inputLimit, inputPage;
 
     public static ClientGetPromotionsFragment newInstance() { return new ClientGetPromotionsFragment(); }
@@ -58,6 +58,7 @@ public class ClientGetPromotionsFragment extends BaseDevFragment {
         inputPage = view.findViewById(R.id.text_page);
         includeMetaBox = view.findViewById(R.id.checkbox_include_meta);
         checkGlobalActivationLimitBox = view.findViewById(R.id.checkbox_check_global_activation_meta);
+        checkBoxincludeVouchers = view.findViewById(R.id.checkbox_include_vouchers);
         view.findViewById(R.id.get_promotions).setOnClickListener(v -> getPromotions());
     }
 
@@ -115,6 +116,7 @@ public class ClientGetPromotionsFragment extends BaseDevFragment {
             query.setSortParameters(sortParams);
             query.setIncludeMeta(includeMetaBox.isChecked());
             query.setCheckGlobalActivationLimits(checkGlobalActivationLimitBox.isChecked());
+            query.setIncludeVouchers(checkBoxincludeVouchers.isChecked());
             apiCall = Promotions.getPromotions(query);
             apiCall.execute(this::onSuccess, new DataActionListener<ApiError>() {
                 @Override